refactor(): divide reflection editor
