+++ ui/apps/akasha/src/extensions/slate-block/slate-editor-block.tsx
@@ -12,7 +12,7 @@ import {
   IEntryData,
   RootExtensionProps,
 } from '@akashaorg/typings/lib/ui';
-import { Draft } from '../inline-editor/utils';
+import { Draft } from '../../utils';
 import { useCreateContentBlockMutation } from '@akashaorg/ui-awf-hooks/lib/generated/hooks-new';
 import {
   AkashaContentBlockBlockDef,
@@ -90,7 +90,7 @@ export const SlateEditorBlock = (
       profileId={'profileId'}
       // actionLabel={''}
       placeholderLabel={'write here'}
-      // disablePublishLabel={disablePublishLabel}
+      // disableActionLabel={disableActionLabel}
       // disablePublish={disablePublish}
       onPublish={() => {
         // void
